requirements.txt
Flask>=3.0.0
pandas>=2.2.0
openpyxl>=3.1.2
Werkzeug>=3.0.1
email-validator>=2.1.0
python-dotenv>=1.0.0
